passing some more tests, working on delete. DB Delete works, network error?

+++ api/routine_activities.js
@@ -1,8 +1,10 @@
 const express = require("express");
+const { getRoutineById } = require("../db");
 const routine_activitiesRouter = express.Router();
 const {
   updateRoutineActivity,
   destroyRoutineActivity,
+  getRoutineActivityById,
 } = require("../db/routine_activities");
 const { requireUser } = require("./utils");
 
@@ -24,14 +26,26 @@ routine_activitiesRouter.patch(
     if (duration) {
       updateFields.duration = duration;
     }
+
     try {
-      // console.log("Jordan", updateFields);
+      // console.log("Jordan!!!!!", updateFields);
+      const rAQuery = await getRoutineActivityById(routineActivityId);
+      const routineQuery = await getRoutineById(rAQuery.routineId);
+      console.log("rAQuery", rAQuery);
+      console.log("routineQuery", routineQuery);
+      console.log(req.user);
+      if (routineQuery.creatorId !== req.user.id) {
+        next();
+      }
       const updatedRA = await updateRoutineActivity(
         routineActivityId,
         updateFields
       );
-      console.log("Pawan test", updatedRA);
+      // if (!updatedRA) {
+      //   next();
+      // } else {
       res.send(updatedRA);
+      // }
     } catch (error) {
       next(error);
     }
@@ -43,8 +57,13 @@ routine_activitiesRouter.delete(
   "/:routineActivityId",
   requireUser,
   async (req, res, next) => {
-    const { routineActivityId } = req.params.routineActivityId;
     try {
+      const rAQuery = await getRoutineActivityById(routineActivityId);
+      const routineQuery = await getRoutineById(rAQuery.routineId);
+      if (routineQuery.creatorId !== req.user.id) {
+        next();
+      }
+      const { routineActivityId } = req.params;
       const destroyed = await destroyRoutineActivity(routineActivityId);
       res.send(destroyed);
     } catch (error) {
